db: do not alias the input buffer after Put
The caller may re-use the slice after storing it in the database, for example
if it came from a buffer they're using to pull in values from somewhere else.
Don't let changes to the caller's buffer reflect in the KV map.

We might also want to do this on the way back out, but this change doesn't do
that (yet).  Alternatively, we might make the input and return values be
strings, which still gives a copy but makes it easier to test.