client: inject controllable poll scheduling for testing
While the production code uses a time.Ticker to schedule polls, it is not
sufficient to plumb in a fake ticker for testing, because the test needs to
know not only when the poll begins, but also when it ends. Shim in an interface
that supersets the ticker to allow tests to regulate the pulse.