client: implement the Store type (#27)
A Store wraps an API client to give the caller a transparent way to get secret
values that automatically stay up-to-date as values change on the server.

While the production code uses a time.Ticker to schedule polls, it is not
sufficient to plumb in a fake ticker for testing, because the test needs to
know not only when the poll begins, but also when it ends. Shim in an interface
that supersets the ticker to allow tests to regulate the pulse.