db: add a dbtest package to simplify testing
The main effect of this change is to pull the unexported testDB wrapper out of
the db_test package into its own utility package. This is a first step toward
adding a similar test-double wrapper for a server, so we can write tests that
call a server without having to reinvent all this.

Update the existing tests to use the new thing.

Signed-off-by: M. J. Fromberger <fromberger@tailscale.com>