setectest: add a package to simplify testing
This package contains a plug-compatible wrapper around server.Server and db.DB
for use with the httptest package to serve secrets in unit tests.

Pulls the unexported testDB wrapper out of the db_test package into an exported
type, and updates the existing tests to use the new thing. Add a Server wrapper
to do the same for the HTTP plumbing, and include some affordances for setup.

Signed-off-by: M. J. Fromberger <fromberger@tailscale.com>