all: report sentinel errors for common response categories
Add sentinel errors to the api package for "not found" and "access denied"
errors, so that users of the client can inspect for these error types.

Make the server encode "not found" errors as a 404 status instead of 500.  We
did not do this initially for concern about information leakage, but since we
do not get this error until we are already willing to vend the requested data
back to the caller (i.e., upon a successful authorization check), we are not
divulging anything they could not have already inferred.

Signed-off-by: M. J. Fromberger <fromberger@tailscale.com>