client/store: add Lookup methods and AllowLookup option (#55)
Some tools have "optional" integrations (e.g., callouts to Slack) that are used
if an API token is available but skipped without error if not, and you can
choose to run with or without it.

To support this, add an AllowLookup option to the Store config. When set, this
permits a Store to start up without a complete list of secrets, and allows the
caller to look up new secrets into the collection on-demand.

Unlike pre-declared secrets, lookups can fail and will not be automatically
retried, but if a lookup succeeds the secret becomes a normal part of the
store's collection.