cmd/setec: convert server address into a flag
Move the server parameter for client subcommands from a positional argument to
a flag, and let the flag default to the $SETEC_SERVER environment variable so
the caller does not need to type it out each time.

Set the flag parser to allow this argument in any position.
Update the command and flax packages to support this.

Examples:

    setec -s https://example.com list
    setec put foo -s https://example.com
    SETEC_SERVER=https://example.com setec info foo