client/setec: check cache validity after decoding
When I changed the cache format, I intended that the old cache would simply
look invalid and be rejected. Because of the way JSON decoding works, though, a
valid JSON value in the wrong shape could result in a cache with invalid
entries.  Notably: An object with expected secret names but the wrong object
shape.

Fix this by explicitly checking for valid secrets after cache loading.