cmd/setec: add a flag to allow "put" of an empty secret value (#87)
Ordinarily an empty secret value is an error, since that is usually not
intended.  In some cases it may be needed, however, e.g., if a program runs in
multiple environments and some of its secrets are not relevant to all of them.

With the --empty-ok flag, "put" will now permit an empty secret.