client/setec: add a StaticTextFile constructor (#92)
A common case we observe in setting up setec is to read a textual API key out
of a file, where leading and trailing spaces should be trimmed from the value
before use. This is common enough that it's worth having its own helper, in
contrast with StaticFile which vends the contents unmodified.