client/setec: document ownership of Secret values and add GetString
Explicitly note that the Secret retains ownership of the bytes it returns, and
document that read-only sharing of the slice is safe.

In addition, add a GetString method to Secret that sugars copying the bytes
into a new string. The caller could also just do this, but having the method
make the intent perhaps more clear.