client/setec: allow a Store to serve secrets entirely locally
Add a StoreClient interface that implements the parts of the setec.Client that
the Store uses. Add a FileClient type that implements this interface, and
modify the Store to accept he interface instead of a concrete Client.

A FileClient uses a local JSON file containing stored secrets, and serves them
locally without using the network. This permits a program using a Store to
start up entirely from local storage, provided the required secrets are saved
locally, even if no setec service is available.

This differs from the cache in that a cache is intended for the case where you
do have a server, and want to work around it being temporarily unavailable.  A
FileClient, by contrast, is intended for bootstrapping or in deployments where
a secrets server is never available.

The operator can construct a secrets file by hand, or by copying the output
written by a FileCache (the file formats are compatible).