client/setec: panic for undeclared secrets when lookups are disabled
We stumbled on a sharp corner, viz., that if you forget to declare a secret to
the Store, the store will start up just fine, but code that refers to that name
will return a nil Secret and panic later at the point of use.

Since a program that does this is already statically incorrect, generate the
panic (with a more helpful diagnostic) immediately when the secret is accessed
rather than letting it show up as a random nil indirection later.

When lookups are enabled, however, retain the current behaviour of reporting a
nil secret or empty Watcher: By enabling lookups the caller has announced that
it is prepared to deal with dynamic conditions, and can be expected to do the
needful.

Add an IsValid method to a Watcher to make this easier to check.