client: unexport the watcher type and update NewUpdater (#118)
The watcher semantics are still useful for plumbing, but it's hard to use
correctly, so unexport it.  Now that watcher is unexported, move its tests from
a separate package into an internal package test.

It is now no longer necessary to have the "watcher" method, so remove it.

An Updater now no longer accepts a watcher (which is unexported), but instead
takes a *Store and the desired secret name directly. Rework the constructor to
allow reporting an error.

Add a StaticUpdater constructor to make an Updater that vends a static value,
analogous to StaticSecret.